backend-api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
python-dotenv
python-jose
passlib
python-multipart
psycopg2-binary
cloudinary
aiofiles
bcrypt
pytest
httpx
pytest-asyncio
asyncpg
shapely
geoalchemy2
pgvector